Promote getAll to TextMapGetter stable API
