+++ context/src/main/java/io/opentelemetry/context/propagation/internal/ExtendedTextMapGetter.java
@@ -6,39 +6,16 @@
 package io.opentelemetry.context.propagation.internal;
 
 import io.opentelemetry.context.propagation.TextMapGetter;
-import java.util.Collections;
-import java.util.Iterator;
-import javax.annotation.Nullable;
 
 /**
- * Extends {@link TextMapGetter} to return possibly multiple values for a given key.
+ * Extended {@link TextMapGetter} with experimental APIs.
  *
  * <p>This class is internal and experimental. Its APIs are unstable and can change at any time. Its
  * APIs (or a version of them) may be promoted to the public stable API in the future, but no
  * guarantees are made.
- *
- * @param <C> carrier of propagation fields, such as an http request.
  */
 public interface ExtendedTextMapGetter<C> extends TextMapGetter<C> {
-  /**
-   * If implemented, returns all values for a given {@code key} in order, or returns an empty list.
-   *
-   * <p>The default method returns the first value of the given propagation {@code key} as a
-   * singleton list, or returns an empty list.
-   *
-   * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
-   * at any time.
-   *
-   * @param carrier carrier of propagation fields, such as an http request.
-   * @param key the key of the field.
-   * @return all values for a given {@code key} in order, or returns an empty list. Default method
-   *     wraps {@code get()} as an {@link Iterator}.
-   */
-  default Iterator<String> getAll(@Nullable C carrier, String key) {
-    String first = get(carrier, key);
-    if (first == null) {
-      return Collections.emptyIterator();
-    }
-    return Collections.singleton(first).iterator();
-  }
+
+  // keep this class even if it is empty, since experimental methods may be added in the future.
+
 }
