+++ context/src/test/java/io/opentelemetry/context/propagation/TextMapGetterTest.java
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.context.propagation.internal;
+package io.opentelemetry.context.propagation;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import com.google.common.collect.ImmutableList;
+import io.opentelemetry.context.propagation.internal.ExtendedTextMapGetter;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Iterator;
@@ -15,9 +16,9 @@ import java.util.List;
 import javax.annotation.Nullable;
 import org.junit.jupiter.api.Test;
 
-class ExtendedTextMapGetterTest {
+class TextMapGetterTest {
 
-  final ExtendedTextMapGetter<Void> nullGet =
+  final TextMapGetter<Void> nullGet =
       new ExtendedTextMapGetter<Void>() {
         @Override
         public Iterable<String> keys(Void carrier) {
@@ -31,7 +32,7 @@ class ExtendedTextMapGetterTest {
         }
       };
 
-  final ExtendedTextMapGetter<Void> nonNullGet =
+  final TextMapGetter<Void> nonNullGet =
       new ExtendedTextMapGetter<Void>() {
         @Override
         public Iterable<String> keys(Void carrier) {
